ai/requirements.txt
flask==3.1.0
openai==1.68.0
python-dotenv==1.1.0
